requirements.txt
aiogram==3.*
SQLAlchemy>=1.4
alembic>=1.9
psycopg2-binary
redis
celery
python-dotenv 